add new fields (stats) to UISoldier model

+++ core/src/main/java/com/rcacao/tactics/core/data/job/model/Job.kt
@@ -8,7 +8,7 @@ data class Job(
     val description: String,
     val move: Int,
     val jump: Int,
-    val pev: Float,
+    val pev: Int,
     val baseStats: BaseStats,
     val cStats: CStats,
     val skills: JobSkill
